+++ c/API/TC_NameFamily_api.h
@@ -1,27 +1,27 @@
 #ifndef TINKERCELL_TC_NAMEANDFAMILY_API_H
 #define TINKERCELL_TC_NAMEANDFAMILY_API_H
 
-#include "../TCstructs.h"
+#include "TCstructs.h"
 BEGIN_C_DECLS
 
 /*! 
  \brief get annotation for this item, i.e. family, author, descriptions, etc.
  \ingroup Annotation
 */
-TCAPIEXPORT ArrayOfStrings tc_getAnnotation(int o);
+TCAPIEXPORT ArrayOfStrings tc_getAnnotation(long o);
 /*! 
  \brief set annotation for this item, i.e. family, author, descriptions, etc.
  \ingroup Annotation
 */
-TCAPIEXPORT void tc_setAnnotation(int o,ArrayOfStrings annot);
+TCAPIEXPORT void tc_setAnnotation(long o,ArrayOfStrings annot);
 
 /*! 
  \brief initialize main
  \ingroup init
 */
 TCAPIEXPORT void tc_NameFamily_api_initialize(
-	   ArrayOfStrings (*tc_getAnnotation0)(int),
-		void (*tc_setAnnotation0)(int,ArrayOfStrings)
+	   ArrayOfStrings (*tc_getAnnotation0)(long),
+		void (*tc_setAnnotation0)(long,ArrayOfStrings)
 	);
 
 END_C_DECLS
